backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.68.2
python-dotenv>=1.0.1